agroshield-backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pillow==10.3.0
opencv-python==4.9.0.80
numpy==1.26.4
ultralytics==8.2.0